Prevent deletion of locations that have at least 1 registered spaceship, updated README

+++ api/controllers/locationController.ts
@@ -105,11 +105,16 @@ export default class LocationController {
       // Check the ID for validity 
       if (this.database.isValidId(req.params.locationID)) {
 
-        // Delete the element with the given ID
         let locationID = parseInt(req.params.locationID);
-        this.database.delete(locationID);
-        res.json({ location_was_deleted: true });
+        let location = this.database.get(locationID);
 
+        // Ensure that no spaceships are at this location before deletion
+        if (location.spaceshipRegistry.size == 0) {
+          this.database.delete(locationID);
+          res.json({ location_was_deleted: true });  
+        } else {
+          res.status(400).json({ location_was_deleted: false, error: "Registered spaceships detected", message: "Cannot delete a location with 1 or more registered spaceships" });
+        }
       } else {
         res.status(400).json({ location_was_deleted: false, error: "Invalid ID", message: "ID does not exist in the system" });
       }
